Fix Store icon Fast Refresh lint
