+++ src/components/StoreProductModal/StoreProductIcon.tsx
@@ -1,22 +1,9 @@
 import type { StoreProductIconName } from '../../vip/vipConfig'
+import { hasStoreProductIcon } from './storeProductIconUtils'
 
 interface StoreProductIconProps {
   name: StoreProductIconName
   className?: string
-}
-
-export const STORE_PRODUCT_ICON_NAMES: readonly StoreProductIconName[] = [
-  'vip',
-  'survivalMode',
-  'publicMode',
-  'tribunalHouse',
-  'dramaMode',
-  'noAds',
-  'fallback',
-]
-
-export function hasStoreProductIcon(name: string): name is StoreProductIconName {
-  return STORE_PRODUCT_ICON_NAMES.includes(name as StoreProductIconName)
 }
 
 function IconArtwork({ name }: { name: StoreProductIconName }) {
